backend/requirements.txt
fastapi
uvicorn[standard]
numpy
opencv-python
pillow
imageio
torch
torchvision
timm
mediapipe
